# docs/crp/groups/orm-raw-repository/baseline/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthenticatedUserResponse": {
        "description": "인증이 돌려주는 사용자(민감 정보 제외).\n\n`user` 기능의 ``UserResponse`` 와 **다른 모델**이다 — 저쪽은 생성·수정 시각까지\n담는 정식 사용자 DTO 고, 이쪽은 로그인·`/me` 가 돌려주는 최소 정보다.\n\n이름을 굳이 다르게 두는 이유는 OpenAPI 때문이다. 서로 다른 모듈의 같은 클래스\n이름은 schema key 가 ``app__features__auth__schemas__auth_schema__UserResponse``\n처럼 **모듈 경로**로 노출된다. 그 이름은 파일을 옮기는 순간 바뀌고, 그때 이\n스키마로 생성한 클라이언트 코드가 통째로 깨진다(DOC-005).",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "is_active"
        ],
        "title": "AuthenticatedUserResponse",
        "type": "object"
      },
      "Body_authLogin": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          },
          "grant_type": {
            "anyOf": [
              {
                "pattern": "^password$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "password": {
            "format": "password",
            "title": "Password",
            "type": "string"
          },
          "scope": {
            "default": "",
            "title": "Scope",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "Body_authLogin",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "API 오류 응답 — 모든 도메인이 같은 모양을 쓴다.\n\n상태 코드마다 다른 모양을 내보내면 클라이언트가 코드별로 분기해야 한다. 하나로\n고정해 두면 오류 처리를 한 곳에 둘 수 있다.\n\n``detail`` 에는 **식별자만** 담는다 — 원본 DB 메시지나 bind 값은 담지 않는다.\n그건 응답으로 나가는 순간 그대로 유출이다(C-12).",
        "examples": [
          {
            "detail": {
              "id": "3f1c9c2e-0b1a-4a5e-9f3d-2c7b8a4d5e6f"
            },
            "error_code": "CATALOG_PRODUCT_NOT_FOUND",
            "message": "상품을 찾을 수 없습니다."
          },
          {
            "detail": {
              "model": "Product"
            },
            "error_code": "DUPLICATE",
            "message": "이미 존재하는 데이터입니다."
          }
        ],
        "properties": {
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "원인을 좁힐 식별자. 값이나 SQL 은 담지 않는다.",
            "examples": [
              {
                "id": "3f1c9c2e-0b1a-4a5e-9f3d-2c7b8a4d5e6f"
              }
            ],
            "title": "Detail"
          },
          "error_code": {
            "description": "기계가 분기할 오류 코드. 도메인별 접두사를 쓴다.",
            "examples": [
              "CATALOG_PRODUCT_NOT_FOUND"
            ],
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "description": "사용자에게 보여줄 한국어 메시지.",
            "examples": [
              "상품을 찾을 수 없습니다."
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProductCreate": {
        "description": "상품 생성 요청.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "상품 설명(선택)",
            "examples": [
              "적축 87키"
            ],
            "title": "Description"
          },
          "is_active": {
            "default": true,
            "description": "판매 여부",
            "examples": [
              true
            ],
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "description": "상품명",
            "examples": [
              "기계식 키보드"
            ],
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "description": "판매가",
            "examples": [
              "129000.00"
            ],
            "title": "Price"
          },
          "sku": {
            "description": "재고 관리 코드(전역 고유)",
            "examples": [
              "SKU-1001"
            ],
            "maxLength": 64,
            "minLength": 1,
            "title": "Sku",
            "type": "string"
          },
          "stock": {
            "default": 0,
            "description": "재고 수량",
            "examples": [
              25
            ],
            "minimum": 0.0,
            "title": "Stock",
            "type": "integer"
          }
        },
        "required": [
          "sku",
          "name",
          "price"
        ],
        "title": "ProductCreate",
        "type": "object"
      },
      "ProductListResponse": {
        "description": "상품 목록 응답(페이지네이션).",
        "examples": [
          {
            "items": [
              {
                "created_at": "2026-08-01T10:30:00+09:00",
                "description": "적축 87키",
                "id": "3f1c9c2e-0b1a-4a5e-9f3d-2c7b8a4d5e6f",
                "is_active": true,
                "name": "기계식 키보드",
                "price": "129000.00",
                "sku": "SKU-1001",
                "stock": 25,
                "updated_at": "2026-08-01T10:30:00+09:00"
              }
            ],
            "limit": 50,
            "skip": 0,
            "total": 1
          }
        ],
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "페이지 크기",
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "skip": {
            "description": "건너뛴 레코드 수",
            "minimum": 0.0,
            "title": "Skip",
            "type": "integer"
          },
          "total": {
            "description": "전체 상품 수",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "ProductListResponse",
        "type": "object"
      },
...
      "ProductUpdate": {
        "description": "상품 수정 요청 — 전달된 필드만 부분 수정한다.\n\n예시가 **일부 필드만** 담고 있는 것이 의도다. 전체를 보내는 모양을 예시로 두면\nPATCH 를 PUT 처럼 쓰게 되고, 빠뜨린 필드가 덮어써진다고 오해하게 된다.",
        "examples": [
          {
            "price": "119000.00",
            "stock": 12
          }
        ],
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "상품 설명",
            "title": "Description"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "판매 여부",
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "상품명",
            "title": "Name"
          },
          "price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "판매가",
            "title": "Price"
          },
          "stock": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "재고 수량",
            "title": "Stock"
          }
        },
        "title": "ProductUpdate",
        "type": "object"
      },
...
      "RefreshRequest": {
        "description": "토큰 재발급 요청.",
        "properties": {
          "refresh_token": {
            "description": "유효한 Refresh Token",
            "title": "Refresh Token",
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "description": "회원가입 요청.",
        "properties": {
          "email": {
            "description": "이메일",
            "maxLength": 255,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "description": "비밀번호(8자 이상)",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "description": "사용자명(고유)",
            "maxLength": 100,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "description": "토큰 응답(OAuth2 bearer).",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n## FastAPI Default Project Structure\n\nRepository 패턴과 Unit of Work 패턴을 적용한 FastAPI 프로젝트 템플릿입니다.\n\n### 주요 기능\n\n- **접속 로그 수집**: 모든 API 요청에 대한 접속 로그 자동 수집\n- **사용자 정보 파싱**: User-Agent 기반 OS, 브라우저, 장치 정보 분석\n- **통계 API**: 장치 유형, OS, 브라우저별 접속 통계\n\n### 아키텍처\n\n```\nRouter → Service → Repository → Database\n           ↑\n      UnitOfWork (트랜잭션 관리)\n```\n\n### 기술 스택\n\n- **FastAPI**: 고성능 비동기 웹 프레임워크\n- **SQLAlchemy 2.0**: 비동기 ORM (aiomysql)\n- **Pydantic v2**: 데이터 검증 및 설정 관리\n- **Scalar**: API 문서 UI\n\n### 환경 설정\n\n| 설정 | 설명 |\n|------|------|\n| `DEBUG=true` | 개발 모드 (DEBUG 로그, 테이블 자동 생성) |\n| `DEBUG=false` | 운영 모드 (INFO 로그, Alembic 마이그레이션 사용) |\n",
    "title": "FastAPI Project",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "description": "username/password(form)로 access·refresh 토큰을 발급합니다.",
        "operationId": "authLogin",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_authLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "로그인(OAuth2 password)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "description": "Bearer access token 으로 현재 로그인 사용자를 반환합니다(보호 엔드포인트).",
        "operationId": "authMe",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "인증 실패"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "현재 사용자",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Refresh Token 으로 새 access·refresh 토큰을 발급합니다.",
        "operationId": "authRefresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "인증 실패"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "토큰 재발급",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "사용자명·이메일·비밀번호로 사용자를 생성합니다.",
        "operationId": "authRegister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "사용자명 중복"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "회원가입",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/catalog/products": {
      "get": {
        "description": "상품 목록을 페이지네이션하여 조회합니다. SKU 오름차순으로 안정 정렬됩니다.",
        "operationId": "listProducts",
        "parameters": [
          {
            "description": "건너뛸 레코드 수(offset)",
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "description": "건너뛸 레코드 수(offset)",
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "description": "조회할 레코드 수(1-100)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "조회할 레코드 수(1-100)",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "판매 중인 상품만 조회",
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": false,
              "description": "판매 중인 상품만 조회",
              "title": "Active Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "상품 목록 조회",
        "tags": [
          "Catalog"
        ]
      },
      "post": {
        "description": "새 상품을 등록합니다. `sku` 는 전역 고유해야 합니다.",
        "operationId": "createProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "이미 존재하는 SKU"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "상품 생성",
        "tags": [
          "Catalog"
        ]
      }
    },
    "/api/v1/catalog/products/{product_id}": {
      "delete": {
        "description": "상품을 삭제합니다.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "description": "상품 ID(UUID)",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "description": "상품 ID(UUID)",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "상품을 찾을 수 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "상품 삭제",
        "tags": [
          "Catalog"
        ]
      },
      "get": {
        "description": "ID 로 상품을 조회합니다.",
        "operationId": "getProduct",
        "parameters": [
          {
            "description": "상품 ID(UUID)",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "description": "상품 ID(UUID)",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "상품을 찾을 수 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "상품 단건 조회",
        "tags": [
          "Catalog"
        ]
      },
      "patch": {
        "description": "상품을 부분 수정합니다(전달한 필드만).",
        "operationId": "updateProduct",
        "parameters": [
          {
            "description": "상품 ID(UUID)",
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "description": "상품 ID(UUID)",
              "title": "Product Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "상품을 찾을 수 없음"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "상품 수정",
        "tags": [
          "Catalog"
        ]
      }
    },
...
    "/api/v1/reports/daily-sales": {
      "get": {
        "description": "기간 내 일자별 주문 수와 매출 합계를 조회합니다. 종료일은 **포함**이며, 결제 완료(`paid`) 주문만 집계합니다.",
        "operationId": "getDailySalesReport",
        "parameters": [
          {
            "description": "조회 시작일(포함)",
            "in": "query",
            "name": "start_date",
            "required": true,
            "schema": {
              "description": "조회 시작일(포함)",
              "examples": [
                "2026-08-01"
              ],
              "format": "date",
              "title": "Start Date",
              "type": "string"
            }
          },
          {
            "description": "조회 종료일(포함)",
            "in": "query",
            "name": "end_date",
            "required": true,
            "schema": {
              "description": "조회 종료일(포함)",
              "examples": [
                "2026-08-07"
              ],
              "format": "date",
              "title": "End Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailySalesReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "조회 기간이 올바르지 않음"
          }
        },
        "summary": "일별 매출 리포트",
        "tags": [
          "Sales Reports"
        ]
      }
    },
...
  },
...
}